+++ src/file_manager/cache_file.hpp
@@ -6,8 +6,8 @@ namespace big
 	class cache_header final
 	{
 	public:
-		uint32_t m_cache_version;
-		uint32_t m_file_version;
+		uint64_t m_cache_version;
+		uint64_t m_file_version;
 		uint64_t m_data_size;
 	};
 
@@ -27,7 +27,7 @@ namespace big
 		/// </summary>
 		/// <param name="cache_file">FileMgr file object</param>
 		/// <param name="cache_version">Internal version, use this to invalidate the cache when changing the structure of the data</param>
-		cache_file(file cache_file, uint32_t cache_version);
+		cache_file(file cache_file, uint64_t cache_version);
 
 		/// <summary>
 		/// Frees any memory used to hold the cached data.
@@ -55,7 +55,7 @@ namespace big
 		/// <param name="game_version">Current Game version</param>
 		/// <param name="online_version">Current Online version</param>
 		/// <returns>True if cache is up to date, false otherwise.</returns>
-		bool up_to_date(uint32_t file_version) const;
+		bool up_to_date(uint64_t file_version) const;
 
 
 		void set_data(cache_data&& data, uint64_t data_size);
@@ -64,14 +64,14 @@ namespace big
 		/// </summary>
 		/// <param name="game_version">Game Build</param>
 		/// <param name="online_version">Online Version</param>
-		void set_header_version(uint32_t file_version);
+		void set_header_version(uint64_t file_version);
 
-		void set_cache_version(uint32_t cache_version);
+		void set_cache_version(uint64_t cache_version);
 
 	private:
 		file m_cache_file;
 
-		uint32_t m_cache_version;
+		uint64_t m_cache_version;
 
 		cache_header m_cache_header;
 		cache_data m_data;
